hopefully this code works
added create, remove and update capabilities to the server

if a piece of data is uploaded twice  and is different (but has the same scouter, replay, match type, etc.), it'll OVERRIDE the old data.
upload a list of auto-generated JSON objects (from dart code).

just read SiteAPI.java for these endpoints (GET and POST)

chrono scouting is a tbd and is not done, its not confirmed.